+++ Booquerio/src/archivos/Registro.h
@@ -16,16 +16,16 @@ class Registro {
 
 public:
   Registro(string unString);
-	Registro(unsigned int id, unsigned short int unaReferencia);
+	Registro(unsigned int id, unsigned int unaReferencia);
 	Registro (string unString, unsigned int id);
 	virtual ~Registro();
 
 	string getString();
 	bool buscarID(unsigned int idBuscado);
 	list<unsigned int>* getIdentificadores();
-	list<unsigned short int>* getReferencias();
+	list<unsigned int>* getReferencias();
 	void agregarId (unsigned int id);
-	void agregarReferencia(unsigned short int);
+	void agregarReferencia(unsigned int);
 
 
 
@@ -34,7 +34,7 @@ private:
 	// Por mas que la referencia sea 1 sola.
 	string miString;
 	list<unsigned int> identificadores;
-	list<unsigned short int> referencias;
+	list<unsigned int> referencias;
 
 };
